tabSRMM 1.0.0.0 "stable" branch
git-svn-id: http://miranda.googlecode.com/svn/branches/stable@3404 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/tabsrmm/chat/chat.h
@@ -370,7 +370,7 @@ extern char *szChatIconString;
 
 #pragma comment(lib,"comctl32.lib")
 
-#define safe_sizeof(a) (sizeof((a)) / sizeof((a)[0]))
+#define safe_sizeof(a) (unsigned int)((sizeof((a)) / sizeof((a)[0])))
 
 #include "chatprototypes.h"
 #include "../msgs.h"
